bashsshrc: exec bash so we don't fork a new shell
When running with SHELL=$SSHHOME/bassshrc we have problems
with certain programs like 'less' and 'mc'.

These programs like 'less' and 'mc' like to interact directly with
the $SHELL process with pipes and wait for pids.  When we fork from
bashsshrc these programs are unable to monitor the subshell and will
hang.

To prevent this and more properly we should exec bash so that we replace
the bashsshrc process with the new shell directly.

Signed-off-by: Ross Brattain <ross.b.brattain@intel.com>